% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_me_docker}
\alias{build_me_docker}
\title{build_me_docker}
\usage{
build_me_docker(
  name = "DockerPlumber",
  version = "rocker/shiny:4.0.1",
  packages = c("shiny"),
  localpath_vector = "app.R"
)
}
\description{
build_me_docker
}
